client

+++ server/src/db/migrations/003-create-category.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       budget_limit: {
         allowNull: false,
-        type: Sequelize.DECIMAL,
+        type: Sequelize.DOUBLE,
       },
       createdAt: {
         type: Sequelize.DATE,
